save downloaded files locally in case we need to reload dataset

+++ utils/data_helpers.py
@@ -16,7 +16,6 @@ from utils.utils import format_file_name, mkdir, normalise, read_pickle
 
 NUM_COLOUR_CHANNELS = 3
 
-
 def _int64_feature(value):
     """Returns an int64_list from a bool / enum / int / uint."""
     if not isinstance(value, list):
@@ -72,6 +71,7 @@ def create_image_caption_tfrecords(tfrecords_dir, image_source_dir, text_source_
         # Read from file and format
         file_names, class_info, text_embeddings = read_text_subset(subset, text_source_dir)
         file_names = [format_file_name(image_source_dir, file_name) for file_name in file_names]
+
         # Convert to bytes
         text_embeddings = [text_embedding.tobytes() for text_embedding in text_embeddings]
         byte_images = get_byte_images(image_paths=file_names, image_dims=image_dims)
@@ -118,9 +118,19 @@ def download_dataset(dataset):
 def download_cub(include_text=False):
     """ Download the birds dataset (CUB-200-2011) """
     BIRDS_DATASET_URL = "http://www.vision.caltech.edu/visipedia-data/CUB-200-2011/CUB_200_2011.tgz"
-    print('Downloading CUB dataset from: {}'.format(BIRDS_DATASET_URL))
-    cub_download_location = pathlib.Path('data/CUB_200_2011.tgz')
-    urllib.request.urlretrieve(BIRDS_DATASET_URL, cub_download_location)
+
+    cub_download_location = 'data/CUB_200_2011.tgz'
+    cub_backup_location = 'data/backup/CUB_200_2011.tgz'
+
+    if os.path.exists(cub_backup_location):
+        print('Retrieving CUB dataset from: {}'.format(cub_backup_location))
+        shutil.copy(cub_backup_location, cub_download_location)
+    else:
+        print('Downloading CUB dataset from: {}'.format(BIRDS_DATASET_URL))
+        cub_download_location = pathlib.Path('data/CUB_200_2011.tgz')
+        urllib.request.urlretrieve(BIRDS_DATASET_URL, cub_download_location)
+        mkdir('data/backup')
+        shutil.copy(cub_download_location, cub_backup_location)
     # Extract image data
     tar = tarfile.open(cub_download_location, "r:gz")
     if include_text:
@@ -134,11 +144,21 @@ def download_cub(include_text=False):
     if include_text:
         # Download the image captions
         BIRDS_TEXT_GDRIVE_ID = '0B3y_msrWZaXLT1BZdVdycDY5TEE'
-        print('Downloading CUB text from Google Drive ID: {}'.format(BIRDS_TEXT_GDRIVE_ID))
+
         cub_text_download_location = "data/birds.zip"
-        gdd.download_file_from_google_drive(file_id=BIRDS_TEXT_GDRIVE_ID,
-                                            dest_path=cub_text_download_location,
-                                            unzip=True)
+        cub_text_backup_location = 'data/backup/birds.zip'
+
+        if os.path.exists(cub_text_backup_location):
+            print('Retrieving CUB dataset from: {}'.format(cub_text_backup_location))
+            shutil.copy(cub_text_backup_location, cub_text_download_location)
+        else:
+            print('Downloading CUB text from Google Drive ID: {}'.format(BIRDS_TEXT_GDRIVE_ID))
+            gdd.download_file_from_google_drive(file_id=BIRDS_TEXT_GDRIVE_ID,
+                                                dest_path=cub_text_download_location,
+                                                unzip=True)
+            mkdir('data/backup')
+            shutil.copy(cub_text_download_location, cub_text_backup_location)
+
         # Move and clean up data
         extracted_text_dir = cub_text_download_location[:-4]
         if os.path.isdir(extracted_text_dir):
@@ -184,21 +204,27 @@ def check_for_xrays(directory):
     shutil.move(f'{train_location}.csv', raw_location)
     shutil.move(f'{valid_location}.csv', raw_location)
 
-def get_byte_images(image_paths, image_dims):
+def get_byte_images(image_paths, image_dims, preprocessing='pad'):
     """ Generate a list of byte representations of each image """
     byte_images_list = []
     for image_path in image_paths:
-        image = Image.open(image_path, 'r')
-        old_size = image.size[:2]
-        ratio = max(image_dims)/max(old_size)
-        new_size = tuple([int(x*ratio) for x in old_size])
-        image = image.resize(new_size, Image.ANTIALIAS)
-        new_img = Image.new('RGB', image_dims)
-        new_img.paste(image, ((image_dims[0]-new_size[0])//2,
-                              (image_dims[1]-new_size[1])//2))
-        img_buffer = io.BytesIO()
-        new_img.save(img_buffer, format='PNG')
-        byte_image = img_buffer.getvalue()
+        if preprocessing == 'pad':
+            image = Image.open(image_path, 'r')
+            old_size = image.size[:2]
+            ratio = max(image_dims)/max(old_size)
+            new_size = tuple([int(x*ratio) for x in old_size])
+            image = image.resize(new_size, Image.ANTIALIAS)
+            new_img = Image.new('RGB', image_dims)
+            new_img.paste(image, ((image_dims[0]-new_size[0])//2,
+                                (image_dims[1]-new_size[1])//2))
+            new_img = new_img * (2/255) - 1
+            img_buffer = io.BytesIO()
+            new_img.save(img_buffer, format='PNG')
+            byte_image = img_buffer.getvalue()
+        elif preprocessing == 'crop':
+            temp = 1
+        else:
+            raise Exception(f"No method available for preprpcessing flag '{preprocessing}' when loading byte images")
         byte_images_list.append(byte_image)
     return byte_images_list
